requirements.txt
discord.py==2.4.0
python-dotenv==1.0.1
pytest==8.3.3
pytest-asyncio==0.24.0
# pymongo==4.8.0 # For MongoDB or 
# psycopg2-binary==2.9.9 # For PostgreSQL
